Trait pallet_exchange::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Currency: Currency<Self::AccountId>;
    type Collection: CollectionInterface<Self::Hash, Self::AccountId>;
    type NFT: NFTInterface<Self::Hash, Self::AccountId>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Currency: Currency<Self::AccountId>

type Collection: CollectionInterface<Self::Hash, Self::AccountId>

type NFT: NFTInterface<Self::Hash, Self::AccountId>

Loading content...